Rename where appropriate by removing v2 postfixes.
