+++ counted_float/_core/benchmarking/__init__.py
@@ -1,6 +1,6 @@
 from importlib.metadata import version
 
-from .flops_v2 import FlopsBenchmarkResults_V2, FlopsBenchmarkSuite_V2
+from .flops import FlopsBenchmarkResults_V2, FlopsBenchmarkSuite
 
 
 def run_flops_benchmark() -> FlopsBenchmarkResults_V2:
@@ -10,7 +10,7 @@ def run_flops_benchmark() -> FlopsBenchmarkResults_V2:
     print(f"Running FLOPS benchmarks using counted-float {version('counted-float')} ...")
     print()
 
-    benchmark_results = FlopsBenchmarkSuite_V2().run()
+    benchmark_results = FlopsBenchmarkSuite().run()
 
     print()
