--- counted_float/_core/benchmarking/flops_v2/__init__.py
@@ -1,2 +1,0 @@
-from ._flops_benchmark_suite_v2 import FlopsBenchmarkResults_V2, FlopsBenchmarkSuite_V2
-from ._flops_micro_benchmark_v2 import FlopsMicroBenchmark_V2
